Django_Institute_Backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
django-environ